+++ src/util.h
@@ -45,9 +45,9 @@ class run_time_ctrl_t
 };
 
 
-std::string botan_aes_cfb_cipher_spec_from_key_byte_len(unsigned key_byte_len);
+std::string botan_aes_cfb_cipher_spec_from_key_byte_len(size_t key_byte_len);
 
-std::string botan_aes_ecb_cipher_spec_from_key_byte_len(unsigned key_byte_len);
+std::string botan_aes_ecb_cipher_spec_from_key_byte_len(size_t key_byte_len);
 
 void lenght_is_multiple_of_aes_block_size_or_throw(std::span<const uint8_t> x);
